cleaned up code, moved log utility functions into new logUtils file
